Add files via upload
The following Test covers the information below.

Write a series of unit tests to verify the features of the Library Management System (LMS) Project.

books can be added to the database
a book can be removed from the database by barcode or title
a book can be checked out.  If so, the due date is no longer "null".
a book can be checked in. If so, the due date is "null".